+++ src/config/axios.js
@@ -4,11 +4,14 @@ import { Auth } from 'aws-amplify';
 axios.interceptors.request.use(async (config) => {
   const authenToken = await Auth.currentSession();
   config.baseURL = 'http://123.31.12.102/kazan/api';
-  console.log("interceptor");
+  config.headers = {
+    'Access-Control-Allow-Origin': 'http://localhost:3000',
+    'Content-type': 'Application/json',
+    'Access-Control-Allow-Methods': 'GET,PUT,POST,DELETE'
+  };
   if (authenToken) {
     config.headers.Authorization = `Bearer ${authenToken.idToken.jwtToken}`;
   }
-  config.headers.AccessControlAllowCors = "*";
-  config.headers.contentType = 'Application/json';
+
   return config
 });
