+++ src/models/index.js
@@ -1,7 +1,5 @@
 export { default as app } from './app'
-// export { default as signUp } from './sign-up'
 export { default as verification } from './verification'
 export { default as login } from './login'
 export { default as register } from './register'
 export { default as group } from './group'
-// export { default as myProfile } from './my-profile'
