+++ src/models/login.js
@@ -1,15 +1,10 @@
-import { Auth } from 'aws-amplify';
-import { dispatch } from '@rematch/core';
+import {Auth} from 'aws-amplify';
+import {dispatch} from '@rematch/core';
 import {push} from "react-router-redux";
-import {SubmissionError} from "redux-form";
 
 export default {
-  state: {
-
-  },
-  reducers: {
-
-  },
+  state: {},
+  reducers: {},
   effects: {
     async loginAsync(payload) {
       const {username, password} = payload;
@@ -18,7 +13,7 @@ export default {
         dispatch.app.isUserLoggedIn(true);
         dispatch.app.setUsername(user.username);
         return dispatch(push('/dashboard'));
-      } catch(err) {
+      } catch (err) {
         if (err.code === 'UserNotConfirmedException') {
           dispatch.verification.setCredential({
             username,
@@ -26,8 +21,8 @@ export default {
           });
           await Auth.resendSignUp(username);
           return dispatch(push('/pages/verification'));
-        } 
-        throw err;        
+        }
+        throw err;
       }
     }
   }
