+++ src/pages/group/index.js
@@ -1,0 +1,2 @@
+import Group from './Group';
+export default Group;
