+++ src/pages/group/table/index.js
@@ -1,0 +1,2 @@
+import GroupTable from './GroupTable';
+export default GroupTable;
