+++ src/pages/home/Home.jsx
@@ -1,5 +1,4 @@
 import React from 'react';
-import axios from 'axios';
 
 export default class Home extends React.PureComponent {
   render() {
