+++ src/pages/register/Register.jsx
@@ -1,14 +1,20 @@
-import React, { PureComponent } from 'react';
-import { connect } from 'react-redux';
-import RegisterForm from './RegisterForm';
+import React from "react";
+import {connect} from "react-redux";
+import RegisterForm from "./RegisterForm";
+import {SubmissionError} from "redux-form";
 
 const mapDispatch = ({ register: { registerAsync }}) => ({
   registerAsync
 });
 
-class Register extends PureComponent {
+class Register extends React.Component {
   handleSubmit = (data) => {
-    return this.props.registerAsync(data);
+    return this.props.registerAsync(data).catch(error => {
+      if (error.message === undefined) {
+        throw new SubmissionError({_error: error})
+      }
+      throw new SubmissionError({_error: error.message});
+    })
   };
 
   render() {
@@ -18,4 +24,4 @@ class Register extends PureComponent {
   }
 }
 
-export default connect(undefined, mapDispatch)(Register);
+export default connect(null, mapDispatch)(Register);
