+++ src/pages/register/index.js
@@ -1,3 +1,2 @@
 import Register from './Register';
-
-export default Register;
+export default Register
