+++ src/pages/register/styled.jsx
