--- src/pages/table-group/TableGroup.jsx
@@ -1,7 +1,0 @@
-import React, {PureComponent} from 'react'
-
-class TableGroup extends PureComponent {
-  render() {
-
-  }
-}
